Wrote mostly-complete test suite for scaler
Added tests for public interface, covering the 
private methods for reading the process list. 
Kept #auto_scale in its own set of tests, because
it's easier that way. May actually make it public,
since the profiler directly calls it. Also 
improved error-handling in several cases, so we
should safely die anytime we can't quite grok
what's going on in the process list.